Split role/directive tests to new file and fix #55
